+++ TrafficIncidentAgent/TrafficIncidentAgentCode/src/main/java/uk/ac/cam/cares/jps/agent/trafficincident/APIConnector.java
@@ -1,4 +1,5 @@
 package uk.ac.cam.cares.jps.agent.trafficincident;
+
 import org.apache.http.client.HttpResponseException;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
@@ -24,7 +25,7 @@ public class APIConnector {
     private String accountKey;
 
     public static final String ERROR_MSG = "APIConnector failed while retrieving readings.";
-    public static final Logger logger = LogManager.getLogger(APIAgentLauncher.class);
+    public static final Logger logger = LogManager.getLogger(TrafficIncidentAgent.class);
 
     public APIConnector(String URL, String date, String key) {
         this.API_URL = URL;
